+++ src/_build/main.py
@@ -13,7 +13,7 @@ ALPHABET="abcdefghijklmnopqrstuvwxyz"
 API_CODE_FILE_PATH="src/sll/api/_generated.c"
 API_HEADER_FILE_PATH="src/include/sll/api/_generated.h"
 COMPILATION_DEFINES=([b"_MSC_VER",b"_WINDOWS",b"WINDLL",b"USERDLL",b"_UNICODE",b"UNICODE"]+([b"NDEBUG"] if "--release" in sys.argv else [b"_DEBUG",b"DEBUG_BUILD"]) if os.name=="nt" else ([] if "--release" in sys.argv else [b"DEBUG_BUILD"]))
-PLATFORM_SOURCE_CODE={"posix":"src/sll/platform/posix.c","nt":"src/sll/platform/windows.c"}
+PLATFORM_SOURCE_CODE={"posix":"src/sll/platform/posix","nt":"src/sll/platform/windows"}
 RETURN_CODE_BASE_TYPE={"0":"I","1":"I","h":"H","Z":"S","f":"F","I":"I","F":"F","C":"C","S":"S","E":"A","A":"A","H":"H","M":"M","O":"O"}
 RETURN_CODE_TYPE_MAP={"0":"return SLL_ACQUIRE_STATIC_INT(0)","1":"return SLL_ACQUIRE_STATIC_INT(1)","h":"return SLL_ACQUIRE_STATIC(handle_zero)","Z":"return SLL_ACQUIRE_STATIC(str_zero)","f":"return SLL_ACQUIRE_STATIC(float_zero)","E":"return SLL_ACQUIRE_STATIC(array_zero)"}
 TYPE_ACCESS_MAP={"I":"$->dt.i","F":"$->dt.f","C":"$->dt.c","S":"&($->dt.s)","A":"&($->dt.a)","H":"&($->dt.h)","M":"&($->dt.m)","O":"$"}
@@ -165,7 +165,12 @@ else:
 		os.environ["LD_LIBRARY_PATH"]="build:"+ld
 if (vb):
 	print("Listing Source Code Files...")
-i_fl=[PLATFORM_SOURCE_CODE[os.name]]
+i_fl=[]
+for r,_,fl in os.walk(PLATFORM_SOURCE_CODE[os.name]):
+	r=r.replace("\\","/").rstrip("/")+"/"
+	for f in fl:
+		if (f[-2:]==".c"):
+			i_fl.append(r+f)
 for r,_,fl in os.walk("src/sll"):
 	r=r.replace("\\","/").rstrip("/")+"/"
 	if ("/platform/" in r.lower()):