+++ src/sll/platform/posix/posix_console.c
@@ -1,0 +1,16 @@
+#include <sll/common.h>
+
+
+
+__SLL_EXTERNAL void sll_platform_enable_console_color(void){
+}
+
+
+
+__SLL_EXTERNAL void sll_platform_reset_console(void){
+}
+
+
+
+__SLL_EXTERNAL void sll_platform_setup_console(void){
+}