+++ src/sll/platform/posix/posix_util.c
@@ -1,0 +1,47 @@
+#include <sll/common.h>
+#include <sll/string.h>
+#include <sll/types.h>
+#include <signal.h>
+#include <sys/types.h>
+#include <time.h>
+#include <unistd.h>
+
+
+
+__SLL_EXTERNAL const sll_char_t* sll_platform_string=SLL_CHAR("posix");
+
+
+
+__SLL_NO_RETURN void _force_exit_platform(void){
+	_exit(SIGABRT);
+}
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_time_t sll_platform_get_current_time(void){
+	struct timespec tm;
+	clock_gettime(CLOCK_REALTIME,&tm);
+	return tm.tv_sec*1000000000+tm.tv_nsec;
+}
+
+
+
+__SLL_EXTERNAL void sll_platform_sleep(sll_time_t tm){
+	struct timespec ts;
+	clock_gettime(CLOCK_REALTIME,&ts);
+	sll_time_t e=ts.tv_sec*1000000000+ts.tv_nsec+tm;
+	while (1){
+		struct timeval tv;
+		tv.tv_sec=tm/1000000000;
+		tv.tv_usec=(tm/1000)%1000000;
+		if (!select(0,NULL,NULL,NULL,&tv)){
+			return;
+		}
+		clock_gettime(CLOCK_REALTIME,&ts);
+		sll_time_t c=ts.tv_sec*1000000000+ts.tv_nsec;
+		if (c>=e){
+			break;
+		}
+		tm=c-e;
+	}
+}